Refactor Java parser to Interface
